requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.10
Flask-Session==0.4.0
psycopg2-binary==2.9.6
itsdangerous==2.1.2
Werkzeug==3.0.4
